smart-email-assistant/backend/requirements.txt
google-auth
google-api-python-client
google-generativeai
fastapi
pandas
beautifulsoup4
python-dotenv
pydantic
uvicorn
